// src/chatbot/enums/chatbot.enum.ts
export enum ChatbotLawyerReqStatusEnum {
  none = 'none',
  requested = 'requested',
  in_process = 'in_process',
  done = 'done',
}

export enum ChatLawyerStatus {
  ACTIVE = 'ACTIVE',
  FINALIZED = 'FINALIZED',
}

export enum ChatTypeEnum {
  CHAT_LAWYER = 'CHAT_LAWYER',
  CHATBOT = 'CHATBOT',
}
export enum HasReqChatbotLawyerEnum {
  yes = 'yes',
  no = 'no',
}